# Conflicts:
#	CMakeLists.txt
#	projects/CMakeLists.txt
#	ui/zenoedit/zenoapplication.h

+++ ui/zenoedit/launch/ztcpserver.cpp
@@ -49,7 +49,7 @@ void ZTcpServer::startProc(const std::string& progJson)
     ZASSERT_EXIT(m_tcpServer);
     if (m_proc && m_proc->isOpen())
     {
-        zeno::log_warn("A program is already running! Please kill first");
+        zeno::log_info("background process already running");
         return;
     }
 
@@ -93,7 +93,7 @@ void ZTcpServer::onNewConnection()
     }
     else
     {
-        zeno::log_info("tcp connection succeed");
+        zeno::log_debug("tcp connection succeed");
 
     }
     connect(m_tcpSocket, SIGNAL(readyRead()), this, SLOT(onReadyRead()));
@@ -129,13 +129,15 @@ void ZTcpServer::onProcFinished(int exitCode, QProcess::ExitStatus exitStatus)
 {
     if (exitStatus == QProcess::NormalExit)
     {
-        m_proc->terminate();
+        if (m_proc)
+            m_proc->terminate();
         m_proc = nullptr;
         zeno::log_info("runner process normally exited with {}", exitCode);
     }
     else if (exitStatus == QProcess::CrashExit)
     {
-        m_proc->terminate();
+        if (m_proc)
+            m_proc->terminate();
         m_proc= nullptr;
         zeno::log_info("runner process crashed with code {}", exitCode);
     }
